[feat] : pakage 이름 변경 및 SecurityConfig 수정
