--- src/main/java/com/example/moa/ProjectMoaApplication.java
@@ -1,13 +1,0 @@
-package com.example.moa;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ProjectMoaApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ProjectMoaApplication.class, args);
-    }
-
-}
