--- src/main/java/com/example/moa/api/jwt/JwtRequestFilter.java
@@ -1,60 +1,0 @@
-package com.example.moa.api.jwt;
-
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.auth0.jwt.exceptions.JWTVerificationException;
-import com.auth0.jwt.exceptions.TokenExpiredException;
-import com.example.moa.user.repository.UserRepository;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-import org.springframework.web.filter.OncePerRequestFilter;
-
-import java.io.IOException;
-
-@RequiredArgsConstructor
-@Component
-@Slf4j
-public class JwtRequestFilter extends OncePerRequestFilter {
-
-    @Autowired
-    UserRepository userRepository;
-
-    @Override
-    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-        String jwtHeader = ((HttpServletRequest)request).getHeader(JwtProperties.HEADER_STRING);
-        System.out.println("JwtRequestFilter 진입");
-
-        // header 가 정상적인 형식인지 확인
-        if(jwtHeader == null || !jwtHeader.startsWith(JwtProperties.TOKEN_PREFIX)) {
-            filterChain.doFilter(request, response);
-            return;
-        }
-
-        // jwt 토큰을 검증해서 정상적인 사용자인지 확인
-        String token = jwtHeader.replace(JwtProperties.TOKEN_PREFIX, "");
-
-        Long userCode = null;
-        //JwtRequestFilter.secret
-        try {
-            userCode = JWT.require(Algorithm.HMAC512(JwtRequestFilter.ALREADY_FILTERED_SUFFIX)).build().verify(token)
-                    .getClaim("id").asLong();
-
-        } catch (TokenExpiredException e) {
-            e.printStackTrace();
-            request.setAttribute(JwtProperties.HEADER_STRING, "토큰이 만료되었습니다.");
-        } catch (JWTVerificationException e) {
-            e.printStackTrace();
-            request.setAttribute(JwtProperties.HEADER_STRING, "유효하지 않은 토큰입니다.");
-        }
-
-        request.setAttribute("userCode", userCode);
-
-        filterChain.doFilter(request, response);
-    }
-}
