--- src/main/java/com/example/moa/api/jwt/dto/OauthToken.java
@@ -1,13 +1,0 @@
-package com.example.moa.api.jwt.dto;
-
-import lombok.Data;
-
-@Data
-public class OauthToken {
-    private String access_token;
-    private String token_type;
-    private String refresh_token;
-    private int expires_in;
-    private String scope;
-    private int refresh_token_expires_in;
-}
