--- src/main/java/com/example/moa/api/jwt/dto/TokenResponse.java
@@ -1,11 +1,0 @@
-package com.example.moa.api.jwt.dto;
-
-
-public record TokenResponse(
-        String grantType,
-        String accessToken,
-        String refreshToken,
-        Long accessTokenExpiresIn
-) {
-
-}
