--- src/main/java/com/example/moa/api/jwt/handler/JwtAccessDeniedHandler.java
@@ -1,24 +1,0 @@
-package com.example.moa.api.jwt.handler;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-
-    @Override
-    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException{
-        // 403에러
-        ObjectMapper objectMapper = new ObjectMapper();
-
-        response.setContentType("application/json;charset=UTF-8");
-        response.setStatus( HttpServletResponse.SC_FORBIDDEN );
-        response.getWriter().write( objectMapper.writeValueAsString("권한이 없습니다.") );
-    }
-}
