--- src/main/java/com/example/moa/api/jwt/oauth/UserInfo.java
@@ -1,125 +1,0 @@
-package com.example.moa.api.jwt.oauth;
-
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.springframework.security.core.CredentialsContainer;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.SpringSecurityCoreVersion;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.util.Assert;
-
-import java.io.Serializable;
-import java.util.*;
-
-
-public class UserInfo implements UserDetails, CredentialsContainer {
-    private static final long serialVersionUID = SpringSecurityCoreVersion.SERIAL_VERSION_UID;
-    private static final Log logger = LogFactory.getLog(User.class);
-    private final String email;
-    private final String name;
-    private final Set<GrantedAuthority> authorities;
-    private final boolean accountNonExpired;
-
-    private final boolean accountNonLocked;
-
-    private final boolean credentialsNonExpired;
-
-    private final boolean enabled;
-
-    public UserInfo(String email, String name, Collection<? extends GrantedAuthority> authorities) {
-        this(email, name, true, true, true, true, authorities);
-    }
-
-    public UserInfo(String email, String name, boolean enabled, boolean accountNonExpired,
-                    boolean credentialsNonExpired, boolean accountNonLocked,
-                    Collection<? extends GrantedAuthority> authorities) {
-        Assert.isTrue(email != null && !"".equals(email) && name != null,
-                "Cannot pass null or empty values to constructor");
-        this.email = email;
-        this.name = name;
-        this.enabled = enabled;
-        this.accountNonExpired = accountNonExpired;
-        this.credentialsNonExpired = credentialsNonExpired;
-        this.accountNonLocked = accountNonLocked;
-        this.authorities = Collections.unmodifiableSet(sortAuthorities(authorities));
-    }
-
-    public String getEmail() {
-        return this.email;
-    }
-
-    public String getName() {
-        return this.name;
-    }
-
-    @Override
-    public Collection<GrantedAuthority> getAuthorities() {
-        return this.authorities;
-    }
-
-    @Override
-    public String getPassword() {
-        return null;
-    }
-
-    @Override
-    public String getUsername() {
-        return null;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return this.enabled;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return this.accountNonExpired;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return this.accountNonLocked;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return this.credentialsNonExpired;
-    }
-
-    @Override
-    public void eraseCredentials() {
-    }
-
-    private static SortedSet<GrantedAuthority> sortAuthorities(Collection<? extends GrantedAuthority> authorities) {
-        Assert.notNull(authorities, "Cannot pass a null GrantedAuthority collection");
-        // Ensure array iteration order is predictable (as per
-        // UserDetails.getAuthorities() contract and SEC-717)
-        SortedSet<GrantedAuthority> sortedAuthorities = new TreeSet<>(new AuthorityComparator());
-        for (GrantedAuthority grantedAuthority : authorities) {
-            Assert.notNull(grantedAuthority, "GrantedAuthority list cannot contain any null elements");
-            sortedAuthorities.add(grantedAuthority);
-        }
-        return sortedAuthorities;
-    }
-
-    private static class AuthorityComparator implements Comparator<GrantedAuthority>, Serializable {
-
-        // @Serial
-        private static final long serialVersionUID = SpringSecurityCoreVersion.SERIAL_VERSION_UID;
-
-        @Override
-        public int compare(GrantedAuthority g1, GrantedAuthority g2) {
-
-            if (g2.getAuthority() == null) {
-                return -1;
-            }
-            if (g1.getAuthority() == null) {
-                return 1;
-            }
-            return g1.getAuthority().compareTo(g2.getAuthority());
-        }
-
-    }
-}
