--- src/main/java/com/example/moa/config/JwtFilterConfig.java
@@ -1,20 +1,0 @@
-package com.example.moa.config;
-
-import com.example.moa.api.jwt.token.TokenProvider;
-import com.example.moa.api.jwt.filter.JwtFilter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@RequiredArgsConstructor
-public class JwtFilterConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-    private final TokenProvider tokenProvider;
-
-    @Override
-    public void configure(HttpSecurity httpSecurity) {
-        JwtFilter customFilter = new JwtFilter(tokenProvider);
-        httpSecurity.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
-    }
-}
