--- src/main/java/com/example/moa/config/SwaggerConfig.java
@@ -1,33 +1,0 @@
-package com.example.moa.config;
-
-import io.swagger.v3.oas.models.Components;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
-import io.swagger.v3.oas.models.security.SecurityRequirement;
-import io.swagger.v3.oas.models.security.SecurityScheme;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpHeaders;
-
-@RequiredArgsConstructor
-@Configuration
-public class SwaggerConfig {
-    @Bean
-    public OpenAPI customOpenAPI() {
-        return new OpenAPI()
-                .info(new Info()
-                        .title("Moa 프로젝트 API")
-                        .description("Moa 백엔드 API 명세서")
-                        .version("v1"))
-                .addSecurityItem(new SecurityRequirement().addList("JWT"))
-                .components(new Components()
-                        .addSecuritySchemes("JWT", new SecurityScheme()
-                                .type(SecurityScheme.Type.HTTP)
-                                .scheme("bearer")
-                                .bearerFormat("JWT")
-                                .in(SecurityScheme.In.HEADER)
-                                .name(HttpHeaders.AUTHORIZATION)));
-    }
-
-}
