--- src/main/java/com/example/moa/user/entity/RoleType.java
@@ -1,13 +1,0 @@
-package com.example.moa.user.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum RoleType {
-    ADMIN("admin"),
-    MEMBER("user");
-
-    private final String value;
-}
