--- src/main/java/com/example/moa/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.example.moa.user.repository;
-
-import com.example.moa.user.entity.UserEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    UserEntity findByEmail(String email);
-}
