+++ src/main/java/com/gdsc/moa/ProjectMoaApplication.java
@@ -1,0 +1,13 @@
+package com.gdsc.moa;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ProjectMoaApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ProjectMoaApplication.class, args);
+    }
+
+}
