+++ src/main/java/com/gdsc/moa/global/config/JwtFilterConfig.java
@@ -1,0 +1,20 @@
+package com.gdsc.moa.global.config;
+
+import com.gdsc.moa.global.jwt.token.TokenProvider;
+import com.gdsc.moa.global.jwt.filter.JwtFilter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.web.DefaultSecurityFilterChain;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+@RequiredArgsConstructor
+public class JwtFilterConfig extends SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
+    private final TokenProvider tokenProvider;
+
+    @Override
+    public void configure(HttpSecurity httpSecurity) {
+        JwtFilter customFilter = new JwtFilter(tokenProvider);
+        httpSecurity.addFilterBefore(customFilter, UsernamePasswordAuthenticationFilter.class);
+    }
+}
