+++ src/main/java/com/gdsc/moa/global/jwt/dto/TokenResponse.java
@@ -1,0 +1,11 @@
+package com.gdsc.moa.global.jwt.dto;
+
+
+public record TokenResponse(
+        String grantType,
+        String accessToken,
+        String refreshToken,
+        Long accessTokenExpiresIn
+) {
+
+}
