+++ src/main/java/com/gdsc/moa/user/entity/RoleType.java
@@ -1,0 +1,13 @@
+package com.gdsc.moa.user.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum RoleType {
+    ADMIN("admin"),
+    MEMBER("user");
+
+    private final String value;
+}
