+++ src/main/java/com/gdsc/moa/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.gdsc.moa.user.repository;
+
+import com.gdsc.moa.user.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+    UserEntity findByEmail(String email);
+}
