+++ src/main/resources/application.yml
@@ -19,7 +19,7 @@ spring:
       enabled: true
       path: /h2-console
   profiles:
-    include: secret
+    include: dev
 
 url:
   base: http://localhost:3000
