--- src/test/java/com/example/moa/ProjectMoaApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.moa;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ProjectMoaApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
