+++ src/test/java/com/gdsc/moa/ProjectMoaApplicationTests.java
@@ -1,0 +1,13 @@
+package com.gdsc.moa;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ProjectMoaApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
